Move add_tokens_to_tokenizer(...) to _shared/ directory.
